+++ src/main/java/kitchenpos/menus/tobe/domain/menu/DisplayedName.java
@@ -1,36 +1,42 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import kitchenpos.shared.domain.Profanities;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Objects;
 
 @Embeddable
-public class MenuGroupName {
+public class DisplayedName {
 
     @Column(name = "name", nullable = false)
     private String name;
 
-    private MenuGroupName() {
+    private DisplayedName() {
     }
 
-    private MenuGroupName(String name) {
-        if (StringUtils.isBlank(name)) {
-            throw new IllegalArgumentException("메뉴 그룹의 이름은 비워 둘 수 없습니다.");
+    private DisplayedName(String name, Profanities profanities) {
+        if (StringUtils.isBlank(name) || profanities.contains(name)) {
+            throw new IllegalArgumentException();
         }
+
         this.name = name;
     }
 
-    public static MenuGroupName of(String name) {
-        return new MenuGroupName(name);
+    public static DisplayedName of(String name, Profanities profanities) {
+        return new DisplayedName(name, profanities);
+    }
+
+    public String getName() {
+        return name;
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        MenuGroupName that = (MenuGroupName) o;
+        DisplayedName that = (DisplayedName) o;
         return Objects.equals(name, that.name);
     }
