+++ src/main/java/kitchenpos/menus/tobe/domain/menu/TobeMenu.java
@@ -1,7 +1,8 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import jakarta.persistence.*;
 import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
 import kitchenpos.products.tobe.domain.Price;
 import kitchenpos.shared.domain.Profanities;
