+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/MenuGroupName.java
@@ -1,42 +1,36 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import kitchenpos.shared.domain.Profanities;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Objects;
 
 @Embeddable
-public class DisplayedName {
+public class MenuGroupName {
 
     @Column(name = "name", nullable = false)
     private String name;
 
-    private DisplayedName() {
+    private MenuGroupName() {
     }
 
-    private DisplayedName(String name, Profanities profanities) {
-        if (StringUtils.isBlank(name) || profanities.contains(name)) {
-            throw new IllegalArgumentException();
+    private MenuGroupName(String name) {
+        if (StringUtils.isBlank(name)) {
+            throw new IllegalArgumentException("메뉴 그룹의 이름은 비워 둘 수 없습니다.");
         }
-
         this.name = name;
     }
 
-    public static DisplayedName of(String name, Profanities profanities) {
-        return new DisplayedName(name, profanities);
-    }
-
-    public String getName() {
-        return name;
+    public static MenuGroupName of(String name) {
+        return new MenuGroupName(name);
     }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        DisplayedName that = (DisplayedName) o;
+        MenuGroupName that = (MenuGroupName) o;
         return Objects.equals(name, that.name);
     }
