+++ src/main/java/kitchenpos/menus/tobe/domain/menugroup/TobeMenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
 import jakarta.persistence.*;
