+++ src/main/java/kitchenpos/menus/tobe/domain/menuproduct/TobeMenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menuproduct;
 
 import jakarta.persistence.*;
