+++ src/test/java/kitchenpos/menus/tobe/domain/menu/TobeMenuProductsTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.menus.domain.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
+import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
