+++ src/test/java/kitchenpos/menus/tobe/domain/menu/TobeMenuTest.java
@@ -1,7 +1,6 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
-import kitchenpos.menus.domain.tobe.domain.TobeMenu;
-import kitchenpos.menus.domain.tobe.domain.TobeMenuProduct;
+import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
 import kitchenpos.share.domain.FakeProfanities;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
