+++ src/test/java/kitchenpos/menus/tobe/domain/menugroup/TobeMenuGroupTest.java
@@ -1,6 +1,5 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menugroup;
 
-import kitchenpos.menus.domain.tobe.domain.TobeMenuGroup;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
